2023/2/13 22:35 实现了关注和关注列表

+++ model/followerModel.go
@@ -4,10 +4,12 @@ import "gorm.io/gorm"
 
 type Follower struct {
 	gorm.Model
-	FollowerID uint64
-	UserID     uint64
+	//粉丝id
+	FollowerID uint
+	//upid
+	UserID uint
 	//冗余字段, 避免多表查询带来的IO损失
 	FollowerName string
-	//是否不再关注user, 默认值设为false,
+	//是否关注user, 默认值设为false, false不关注, true关注
 	IsFollow bool `gorm:"default:false"`
 }
